Reorganize test utility traits into Unit/Concerns directory
