+++ tests/Unit/Concerns/ArraySubsetAsserts.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rareloop\Lumberjack\Test\Unit;
+namespace Rareloop\Lumberjack\Test\Unit\Concerns;
 
 use PHPUnit\Framework\Assert;
