+++ tests/Unit/Concerns/BrainMonkeyPHPUnitIntegration.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rareloop\Lumberjack\Test\Unit;
+namespace Rareloop\Lumberjack\Test\Unit\Concerns;
 
 use Brain\Monkey;
