+++ tests/Unit/Providers/SessionServiceProviderTest.php
@@ -7,7 +7,7 @@ use Rareloop\Lumberjack\Test\TestCase;
 use Rareloop\Lumberjack\Application;
 use Rareloop\Lumberjack\Providers\SessionServiceProvider;
 use Rareloop\Lumberjack\Session\SessionManager;
-use Rareloop\Lumberjack\Test\Unit\BrainMonkeyPHPUnitIntegration;
+use Rareloop\Lumberjack\Test\Unit\Concerns\BrainMonkeyPHPUnitIntegration;
 
 class SessionServiceProviderTest extends TestCase
 {
