+++ tests/Unit/Providers/ThemeSupportServiceProviderTest.php
@@ -9,7 +9,7 @@ use Rareloop\Lumberjack\Test\TestCase;
 use Rareloop\Lumberjack\Application;
 use Rareloop\Lumberjack\Config;
 use Rareloop\Lumberjack\Providers\ThemeSupportServiceProvider;
-use Rareloop\Lumberjack\Test\Unit\BrainMonkeyPHPUnitIntegration;
+use Rareloop\Lumberjack\Test\Unit\Concerns\BrainMonkeyPHPUnitIntegration;
 
 class ThemeSupportServiceProviderTest extends TestCase
 {
