--- tests/Unit/RestoresHandlers.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Rareloop\Lumberjack\Test\Unit;
-
-trait RestoresHandlers
-{
-    protected function restoreHandlers(): void
-    {
-        restore_error_handler();
-        restore_exception_handler();
-    }
-}
